refactor(domain): refactor update imports after changing domain package name to model (#11)
