+++ src/main/java/org/instituteatri/deep/model/token/TokenType.java
@@ -1,4 +1,4 @@
-package org.instituteatri.deep.domain.token;
+package org.instituteatri.deep.model.token;
 
 public enum TokenType {
     BEARER
